feat(speedup): add basic plot implementation which can read data from csr and plot speedup figure
This commit add a performance plot implementation. It now read a csv file as perforamcne data and
allow user to selection the column as algorithm data or x/y axis data. After all data are confirmed
and "submit" button is click, the speedup plot and the raw data can be presented.