feat(speedup): add segmented speedup statistics
In segmented speedup statistics, we can split the x axis into multiple ranges. For each range, we
can calculate the speedup of the two compared algorithms and list the results in a table. By default
(no segmentation is given), there is only one range (which is the whole x axis).